+++ pyviews/binding/tests/common.py
@@ -1,4 +1,4 @@
-from pyviews.core.bindable import BindableEntity
+from pyviews.core.binding import BindableEntity
 
 
 class InnerViewModel(BindableEntity):
